Implemented spinlock using atomic increment/decrement assembly routines.

+++ ValvanoWareTM4C123/lab2/os.h
@@ -34,7 +34,7 @@
 
 // feel free to change the type of semaphore, there are lots of good solutions
 struct  Sema4{
-  long Value;   // >0 means free, otherwise means busy        
+  int32_t Value;   // >0 means free, otherwise means busy
 // add other components here, if necessary to implement blocking
 };
 typedef struct Sema4 Sema4Type;
